requirements.txt
geopandas
libpysal 
esda